Enable ios unit tests for amplify_flutter in circleci
