+++ packages/amplify_flutter/example/ios/unit_tests/Info.plist
