+++ packages/amplify_flutter/ios/amplify_flutter.podspec
@@ -1,6 +1,6 @@
 #
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html.
-# Run `pod lib lint amplify.podspec' to validate before publishing.
+# Run `pod lib lint amplify_flutter.podspec' to validate before publishing.
 #
 Pod::Spec.new do |s|
   s.name             = 'amplify_flutter'
